linux 64.

+++ src/org/eclipse/swt/dnd/TableDragSourceEffect.d
@@ -103,7 +103,7 @@ public class TableDragSourceEffect : DragSourceEffect {
         void* model = OS.GTK_VERSION < OS.buildVERSION (2, 2, 4) ? &dummy : null;
         auto list = OS.gtk_tree_selection_get_selected_rows (selection, &model);
         if (list is null) return null;
-        int count = Math.min(10, OS.g_list_length (list));
+        ptrdiff_t count = Math.min(10, OS.g_list_length (list));
 
         Display display = table.getDisplay();
         if (count is 1) {
